+++ meta/CHANGELOG.md
@@ -2,8 +2,19 @@
 
 All notable changes to the Orbital Trading project will be documented in this file by the user, only. Gemini is never to modify or provide updates to this file, and must always treat this file as read-only. The user will always manually update this file.
 
+## [33.60] - 2026-1-2
+-  added Refined Propellant assets
+
+## [33.59] - 2026-1-1
+-  updated AssetService to handle 0 assets
+-  updated assets_config file names
+-  improved contrast of commodity cards
+-  added Plasteel assets
+-  added Hydroponics assets
+-  added Cybernetics assets
+
 ## [33.56] - 2026-1-1
--  updated Water Ice assets (5)
+-  updated Water Ice assets (1)
 -  added commodity card background functionality
 -  added commodity card asset cycling
 -  updated general asset cycling each day